:recycle: [#4686] Refactor duplicated code
Created a generic component for the button-that-opens-modal with error
icon next to it, which renders the plugin-specific configuration form
inside. This puts the less elegant code in a nice isolated box instead
of spreading across the code base.
