+++ src/openforms/js/components/admin/form_design/registrations/objectsapi/ObjectsApiOptionsForm.js
@@ -1,21 +1,14 @@
-import {Formik} from 'formik';
 import PropTypes from 'prop-types';
-import React, {useContext, useState} from 'react';
-import {FormattedMessage, useIntl} from 'react-intl';
+import React, {useContext} from 'react';
+import {FormattedMessage} from 'react-intl';
 
-import {SubmitAction} from 'components/admin/forms/ActionButton';
-import Field from 'components/admin/forms/Field';
-import SubmitRow from 'components/admin/forms/SubmitRow';
+import OptionsConfiguration from 'components/admin/form_design/registrations/shared/OptionsConfiguration';
 import {ValidationErrorContext} from 'components/admin/forms/ValidationErrors';
-import {ErrorIcon} from 'components/admin/icons';
-import {FormModal} from 'components/admin/modals';
 
 import ObjectsApiOptionsFormFields from './ObjectsApiOptionsFormFields';
 import {filterErrors, getChoicesFromSchema} from './utils';
 
 const ObjectsApiOptionsForm = ({index, name, label, schema, formData, onChange}) => {
-  const intl = useIntl();
-  const [modalOpen, setModalOpen] = useState(false);
   const validationErrors = useContext(ValidationErrorContext);
   const {objectsApiGroup} = schema.properties;
   const apiGroupChoices = getChoicesFromSchema(objectsApiGroup.enum, objectsApiGroup.enumNames);
@@ -23,90 +16,25 @@ const ObjectsApiOptionsForm = ({index, name, label, schema, formData, onChange})
   const defaultGroup = apiGroupChoices.length === 1 ? apiGroupChoices[0][0] : undefined;
 
   return (
-    <Field name={name} label={label}>
-      <>
-        <span style={{display: 'inline-flex', gap: '10px', alignItems: 'center'}}>
-          <button
-            type="button"
-            className="button"
-            onClick={e => {
-              e.preventDefault();
-              setModalOpen(true);
-            }}
-            // admin style overrides...
-            style={{
-              paddingInline: '15px',
-              paddingBlock: '10px',
-            }}
-          >
-            <FormattedMessage
-              description="Link label to open registration options modal"
-              defaultMessage="Configure options"
-            />
-          </button>
-
-          {numErrors > 0 && (
-            <ErrorIcon
-              text={intl.formatMessage(
-                {
-                  description: 'Objects API registration validation errors icon next to button',
-                  defaultMessage: `{numErrors, plural,
-              one {There is a validation error.}
-              other {There are {numErrors} validation errors.}
-            }`,
-                },
-                {numErrors}
-              )}
-              extraClassname="fa-lg icon icon--danger icon--compact icon--no-pointer"
-            />
-          )}
-        </span>
-
-        <FormModal
-          isOpen={modalOpen}
-          title={
-            <FormattedMessage
-              description="Objects API registration options modal title"
-              defaultMessage="Plugin configuration: Objects API"
-            />
-          }
-          closeModal={() => setModalOpen(false)}
-          extraModifiers={['large']}
-        >
-          <Formik
-            initialValues={{
-              ...formData,
-              // Ensure that if there's only one option, it is automatically selected.
-              objectsApiGroup: formData.objectsApiGroup ?? defaultGroup,
-            }}
-            onSubmit={(values, actions) => {
-              onChange({formData: values});
-              actions.setSubmitting(false);
-              setModalOpen(false);
-            }}
-          >
-            {({handleSubmit}) => (
-              <>
-                <ObjectsApiOptionsFormFields
-                  index={index}
-                  name={name}
-                  apiGroupChoices={apiGroupChoices}
-                />
-
-                <SubmitRow>
-                  <SubmitAction
-                    onClick={event => {
-                      event.preventDefault();
-                      handleSubmit(event);
-                    }}
-                  />
-                </SubmitRow>
-              </>
-            )}
-          </Formik>
-        </FormModal>
-      </>
-    </Field>
+    <OptionsConfiguration
+      name={name}
+      label={label}
+      numErrors={numErrors}
+      modalTitle={
+        <FormattedMessage
+          description="Objects API registration options modal title"
+          defaultMessage="Plugin configuration: Objects API"
+        />
+      }
+      initialFormData={{
+        ...formData,
+        // Ensure that if there's only one option, it is automatically selected.
+        objectsApiGroup: formData.objectsApiGroup ?? defaultGroup,
+      }}
+      onSubmit={values => onChange({formData: values})}
+    >
+      <ObjectsApiOptionsFormFields index={index} name={name} apiGroupChoices={apiGroupChoices} />
+    </OptionsConfiguration>
   );
 };
