+++ src/openforms/js/components/admin/form_design/registrations/shared/OptionsConfiguration.js
@@ -1,0 +1,107 @@
+import {Formik} from 'formik';
+import PropTypes from 'prop-types';
+import {useState} from 'react';
+import {FormattedMessage, useIntl} from 'react-intl';
+
+import {SubmitAction} from 'components/admin/forms/ActionButton';
+import Field from 'components/admin/forms/Field';
+import SubmitRow from 'components/admin/forms/SubmitRow';
+import {ErrorIcon} from 'components/admin/icons';
+import {FormModal} from 'components/admin/modals';
+
+const OptionsConfiguration = ({
+  name,
+  label,
+  numErrors,
+  modalTitle,
+  initialFormData,
+  onSubmit,
+  children,
+}) => {
+  const intl = useIntl();
+  const [modalOpen, setModalOpen] = useState(false);
+  return (
+    <Field name={name} label={label}>
+      <>
+        <span style={{display: 'inline-flex', gap: '10px', alignItems: 'center'}}>
+          <button
+            type="button"
+            className="button"
+            onClick={e => {
+              e.preventDefault();
+              setModalOpen(true);
+            }}
+            // admin style overrides...
+            style={{
+              paddingInline: '15px',
+              paddingBlock: '10px',
+            }}
+          >
+            <FormattedMessage
+              description="Link label to open registration options modal"
+              defaultMessage="Configure options"
+            />
+          </button>
+          {numErrors > 0 && (
+            <ErrorIcon
+              text={intl.formatMessage(
+                {
+                  description:
+                    'Registration validation errors icon next to button to configure options',
+                  defaultMessage: `{numErrors, plural,
+            one {There is a validation error.}
+            other {There are {numErrors} validation errors.}
+          }`,
+                },
+                {numErrors}
+              )}
+              extraClassname="fa-lg icon icon--danger icon--compact icon--no-pointer"
+            />
+          )}
+        </span>
+
+        <FormModal
+          isOpen={modalOpen}
+          title={modalTitle}
+          closeModal={() => setModalOpen(false)}
+          extraModifiers={['large']}
+        >
+          <Formik
+            initialValues={initialFormData}
+            onSubmit={(values, actions) => {
+              onSubmit(values);
+              actions.setSubmitting(false);
+              setModalOpen(false);
+            }}
+          >
+            {({handleSubmit}) => (
+              <>
+                {children}
+
+                <SubmitRow>
+                  <SubmitAction
+                    onClick={event => {
+                      event.preventDefault();
+                      handleSubmit(event);
+                    }}
+                  />
+                </SubmitRow>
+              </>
+            )}
+          </Formik>
+        </FormModal>
+      </>
+    </Field>
+  );
+};
+
+OptionsConfiguration.propTypes = {
+  name: PropTypes.string.isRequired,
+  label: PropTypes.node.isRequired,
+  modalTitle: PropTypes.node.isRequired,
+  numErrors: PropTypes.number.isRequired,
+  initialFormData: PropTypes.object.isRequired,
+  onSubmit: PropTypes.func.isRequired,
+};
+
+export default OptionsConfiguration;
