+++ src/openforms/js/components/admin/form_design/registrations/zgw/ZGWOptionsForm.js
@@ -1,21 +1,14 @@
-import {Formik} from 'formik';
 import PropTypes from 'prop-types';
-import React, {useContext, useState} from 'react';
-import {FormattedMessage, useIntl} from 'react-intl';
+import React, {useContext} from 'react';
+import {FormattedMessage} from 'react-intl';
 
-import {SubmitAction} from 'components/admin/forms/ActionButton';
-import Field from 'components/admin/forms/Field';
-import SubmitRow from 'components/admin/forms/SubmitRow';
+import OptionsConfiguration from 'components/admin/form_design/registrations/shared/OptionsConfiguration';
 import {ValidationErrorContext} from 'components/admin/forms/ValidationErrors';
-import {ErrorIcon} from 'components/admin/icons';
-import {FormModal} from 'components/admin/modals';
 
 import ZGWFormFields from './ZGWOptionsFormFields';
 import {filterErrors, getChoicesFromSchema} from './utils';
 
-const ZGWOptionsForm = ({index, name, label, schema, formData, onChange}) => {
-  const intl = useIntl();
-  const [modalOpen, setModalOpen] = useState(false);
+const ZGWOptionsForm = ({name, label, schema, formData, onChange}) => {
   const validationErrors = useContext(ValidationErrorContext);
 
   const {zgwApiGroup, zaakVertrouwelijkheidaanduiding} = schema.properties;
@@ -27,105 +20,44 @@ const ZGWOptionsForm = ({index, name, label, schema, formData, onChange}) => {
 
   const numErrors = filterErrors(name, validationErrors).length;
   const defaultGroup = apiGroupChoices.length === 1 ? apiGroupChoices[0][0] : undefined;
+
   return (
-    <Field name={name} label={label}>
-      <>
-        <span style={{display: 'inline-flex', gap: '10px', alignItems: 'center'}}>
-          <button
-            type="button"
-            className="button"
-            onClick={e => {
-              e.preventDefault();
-              setModalOpen(true);
-            }}
-            // admin style overrides...
-            style={{
-              paddingInline: '15px',
-              paddingBlock: '10px',
-            }}
-          >
-            <FormattedMessage
-              description="Link label to open registration options modal"
-              defaultMessage="Configure options"
-            />
-          </button>
-
-          {numErrors > 0 && (
-            <ErrorIcon
-              text={intl.formatMessage(
-                {
-                  description: 'Objects API registration validation errors icon next to button',
-                  defaultMessage: `{numErrors, plural,
-              one {There is a validation error.}
-              other {There are {numErrors} validation errors.}
-            }`,
-                },
-                {numErrors}
-              )}
-              extraClassname="fa-lg icon icon--danger icon--compact icon--no-pointer"
-            />
-          )}
-        </span>
-
-        <FormModal
-          isOpen={modalOpen}
-          title={
-            <FormattedMessage
-              description="ZGW APIs registration options modal title"
-              defaultMessage="Plugin configuration: ZGW APIs"
-            />
-          }
-          closeModal={() => setModalOpen(false)}
-          extraModifiers={['large']}
-        >
-          <Formik
-            initialValues={{
-              // defaults
-              caseTypeIdentification: '',
-              zaaktype: '',
-              informatieobjecttype: '',
-              organisatieRsin: '',
-              zaakVertrouwelijkheidaanduiding: '',
-              medewerkerRoltype: '',
-              propertyMappings: [],
-              // saved data, overwrites defaults
-              ...formData,
-              // Ensure that if there's only one option, it is automatically selected.
-              zgwApiGroup: formData.zgwApiGroup ?? defaultGroup,
-            }}
-            onSubmit={(values, actions) => {
-              onChange({formData: values});
-              actions.setSubmitting(false);
-              setModalOpen(false);
-            }}
-          >
-            {({handleSubmit}) => (
-              <>
-                <ZGWFormFields
-                  name={name}
-                  apiGroupChoices={apiGroupChoices}
-                  confidentialityLevelChoices={confidentialityLevelChoices}
-                />
-
-                <SubmitRow>
-                  <SubmitAction
-                    onClick={event => {
-                      event.preventDefault();
-                      handleSubmit(event);
-                    }}
-                  />
-                </SubmitRow>
-              </>
-            )}
-          </Formik>
-        </FormModal>
-      </>
-    </Field>
+    <OptionsConfiguration
+      name={name}
+      label={label}
+      numErrors={numErrors}
+      modalTitle={
+        <FormattedMessage
+          description="ZGW APIs registration options modal title"
+          defaultMessage="Plugin configuration: ZGW APIs"
+        />
+      }
+      initialFormData={{
+        // defaults
+        caseTypeIdentification: '',
+        zaaktype: '',
+        informatieobjecttype: '',
+        organisatieRsin: '',
+        zaakVertrouwelijkheidaanduiding: '',
+        medewerkerRoltype: '',
+        propertyMappings: [],
+        // saved data, overwrites defaults
+        ...formData,
+        // Ensure that if there's only one option, it is automatically selected.
+        zgwApiGroup: formData.zgwApiGroup ?? defaultGroup,
+      }}
+      onSubmit={values => onChange({formData: values})}
+    >
+      <ZGWFormFields
+        name={name}
+        apiGroupChoices={apiGroupChoices}
+        confidentialityLevelChoices={confidentialityLevelChoices}
+      />
+    </OptionsConfiguration>
   );
 };
 
 ZGWOptionsForm.propTypes = {
-  index: PropTypes.number,
   name: PropTypes.string,
   label: PropTypes.node,
   schema: PropTypes.shape({
